Move to stride_tricks
