+++ dask/array/tests/test_overlap.py
@@ -789,7 +789,7 @@ def test_ensure_minimum_chunksize_raises_error():
 @pytest.mark.parametrize("window_shape, axis", [((3, 2), (1, 2))])
 def test_sliding_window_view(window_shape, axis):
     from ..numpy_compat import sliding_window_view as np_sliding_window_view
-    from ..overlap import sliding_window_view
+    from ..lib.stride_tricks import sliding_window_view
 
     arr = da.from_array(np.arange(60.0).reshape(3, 4, 5), chunks=(2, 3, 4))
     actual = sliding_window_view(arr, window_shape, axis)
