+++ docs/source/array-api.rst
@@ -355,7 +355,7 @@ Slightly Overlapping Computations
 .. autosummary::
    overlap.overlap
    overlap.map_overlap
-   overlap.sliding_window_view
+   lib.stride_tricks.sliding_window_view
    overlap.trim_internal
    overlap.trim_overlap
